More hacks.
Give scroll control back to the browser.  On mobile this feels quite nice. On desktop,
less so and it means you can't scroll in 2 directions at once.
But I've hackily avoided processing any touch events on the workspace and let them
be handled by the browser instead.

There are lots of broken things:
- the workspace is no longer infinite. That's just hacked to be really big.
- dropdowns and whatnot are no longer hidden on mouse down so if you open one up and
scroll its all kinds of broken
- I didn't reposition the trash/zoom buttons so they're way the heck out there.
- the toolbox isn't really positioned correctly either.  But it mostly works.
- and probably many other things.