NOTE. Previous commit message incorrect. THis commit does:
Set flyout's scrollbar visiblity to false when the flyout is hidden. This used to work
because the scrollbar was contained in the svg that is getting display none'd. Now
it is a sibbling so it needs to be set explicitly.

The previous commit with that description (d3a53c543b1a0529447b45d1c4fef826e0266b62) actually
positioned the flyout relative to the toolbox rather than underneath it.